chore(pr) Move all the deep equality work out of /core/shared.
