+++ editor/src/utils/deep-equality.spec.ts
